typify
